requirements.txt
Flask==2.0.3
twilio==6.63.0
Pillow==8.4.0
